refactor: StampPaletteAddStamp -> StampPaletteEditorAddStamp
