+++ src/components/Settings/StampPaletteTab/StampPaletteEditorAddStamp.vue
